feat/10 게시물 조회 구현 및 인가 개념 적용

+++ src/main/java/com/hanghae/newsfeed/post/service/PostService.java
@@ -4,12 +4,17 @@ import com.hanghae.newsfeed.post.dto.request.PostRequestDto;
 import com.hanghae.newsfeed.post.dto.response.PostResponseDto;
 import com.hanghae.newsfeed.post.entity.Post;
 import com.hanghae.newsfeed.post.repository.PostRepository;
+import com.hanghae.newsfeed.security.UserDetailsImpl;
 import com.hanghae.newsfeed.user.entity.User;
 import com.hanghae.newsfeed.user.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Slf4j
 @Service
@@ -18,8 +23,27 @@ public class PostService {
     private final PostRepository postRepository;
     private final UserRepository userRepository;
 
+    // 게시물 목록 조회
+    public List<PostResponseDto> getAllPosts() {
+        List<Post> allPosts = postRepository.findAll();
+
+        return allPosts.stream()
+                .map(post -> PostResponseDto.createPostDto(post, "게시물 조회 성공"))
+                .collect(Collectors.toList());
+    }
+
+    // 게시물 조회
+    public PostResponseDto getPost(Long postId) {
+        // 게시물 조회 예외 발생
+        Post post = postRepository.findById(postId)
+                .orElseThrow(() -> new IllegalArgumentException("등록된 게시물이 없습니다."));
+
+        return PostResponseDto.createPostDto(post, "게시물 조회 성공");
+    }
+
+    // 게시물 작성
     @Transactional
-    public PostResponseDto create(PostRequestDto requestDto) {
+    public PostResponseDto createPost(PostRequestDto requestDto) {
         // 유저 조회 예외 발생
         User user = userRepository.findById(requestDto.getUserId())
                     .orElseThrow(() -> new IllegalArgumentException("게시물 작성 실패, 등록된 사용자가 없습니다."));
@@ -32,11 +56,17 @@ public class PostService {
         return PostResponseDto.createPostDto(createdPost, "게시물 작성 성공");
     }
 
+    // 게시물 수정
     @Transactional
-    public PostResponseDto update(Long id, PostRequestDto requestDto) {
+    public PostResponseDto updatePost(Long postId, PostRequestDto requestDto, UserDetailsImpl userDetails) {
         // 게시물 조회 예외 발생
-        Post target = postRepository.findById(id)
+        Post target = postRepository.findById(postId)
                 .orElseThrow(() -> new IllegalArgumentException("게시물 수정 실패, 등록된 게시물이 없습니다."));
+
+        // 게시물 작성자와 현재 로그인한 사용자의 일치 여부 확인 && 관리자 확인
+        if (!target.getUser().getId().equals(userDetails.getId()) && !userDetails.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN"))) {
+            throw new IllegalStateException("게시물 수정 권한이 없습니다.");
+        }
 
         // 게시물 수정
         target.patch(requestDto);
@@ -47,11 +77,17 @@ public class PostService {
         return PostResponseDto.createPostDto(updatedPost, "게시물 수정 성공");
     }
 
+    // 게시물 삭제
     @Transactional
-    public PostResponseDto delete(Long id) {
+    public PostResponseDto deletePost(Long postId, UserDetailsImpl userDetails) {
         // 게시물 조회 예외 발생
-        Post target = postRepository.findById(id)
+        Post target = postRepository.findById(postId)
                 .orElseThrow(() -> new IllegalArgumentException("게시물 삭제 실패, 등록된 게시물이 없습니다."));
+
+        // 게시물 작성자와 현재 로그인한 사용자의 일치 여부 확인\
+        if (!target.getUser().getId().equals(userDetails.getId()) && !userDetails.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN"))) {
+            throw new IllegalStateException("게시물 삭제 권한이 없습니다.");
+        }
         
         // 게시물 삭제
         postRepository.delete(target);
